Refactor reconcileaction to have a message
This refactors the reconcileaction package to also return a string
message as part of the result. The intent is to use this to explain the
True/False/Unknown result of the action. It will eventually be placed in
the CR status field along with the list of ConditionStatus.

As a part of this change, ReconcileAction was renamed to Action since
it's in the reconcileaction package, and a Result struct was added to
carry both the condition and the message.

Signed-off-by: John Strunk <jstrunk@redhat.com>